+++ Changelog.md
@@ -8,6 +8,8 @@ Deprecations:
 * Deprecate the `host` argument of `RSpec::Mocks.setup`. Instead
   `RSpec::Mocks::ExampleMethods` should be included directly in the scope where
   RSpec's mocking capabilities are used. (Sam Phippen)
+* Deprecate using any of rspec-mocks' features outside the per-test
+  lifecycle (e.g. from a `before(:all)` hook). (Myron Marston)
 
 ### 2.99.0.beta1 / 2013-11-07
 [full changelog](http://github.com/rspec/rspec-mocks/compare/v2.14.4...v2.99.0.beta1)